Move GSObjectModel's reflux implementation to Model, since it's generic (so now works for Optical
